Feat: 로그아웃 API 접근 권한 및 로직 수정

+++ wtnt/user/auth/service.py
@@ -55,12 +55,16 @@ class AuthService(BaseService):
         return response_data, access_token
 
     def logout(self):
-        _, access_token = self.request.META.get("HTTP_AUTHORIZATION").split(" ")
-        try:
-            user_id = AccessToken(access_token, verify=False).payload.get("user_id")
-        except TokenError:
-            raise token_exception.InvalidTokenError()
-        RedisUtils.delete_refresh_token(user_id)
+        authorization = self.request.META.get("HTTP_AUTHORIZATION", None)
+        if authorization:
+            _, access_token = authorization.split(" ")
+            try:
+                user_id = AccessToken(access_token, verify=False).payload.get("user_id")
+            except TokenError:
+                raise token_exception.InvalidTokenError()
+            RedisUtils.delete_refresh_token(user_id)
+        else:
+            token_exception.NoTokenInAuthorizationHeaderError()
 
 
 class RegisterService(BaseService):
